transpiler: long returns are INTEGER; `i` members and by-ref; this. only to disambiguate
Three follow-ups to the `i` prefix and one request from Alex, taken in
one rebuild of both targets.

hbfuncs.tab records a C# integral return as INTEGER (genfunctab.py; the
five hb_bit* functions and StrCmp moved from NUMERIC). A bit setter,
`RETURN IIF( lx, hb_bitOr(...), hb_bitAnd(...) )`, now returns long, so
the 100 in settflag.prg and buffflag.prg write into AS INTEGER members
(oLine:nFlags) with no cast: ORM-NARROW 282 -> 210 in easipos.

A call or method returning long is integral to the emitter
(hb_csExprIsCsIntegral asks hb_csExprCsType). That serves two things.
A division of one gets the (decimal) that keeps `/` Harbour's: C#'s
`long / int` is integer division, and GetFlagI(), long since 7535e0d888,
was already divided that way at easiudm.prg (harmless there, under
Int()). And hb_csNeedsIntCast drops the (long) it put around a value
already long: 237 casts gone in easipos, 236 in testeasipos.

test121 covers the `i` cases test120 did not, and they turned up four
gaps:
- `VAR iCount INIT 0` was decimal: a member's INIT beat its prefix
  (hb_astInferTypeFromInit now applies the `i` rule a local's
  initializer has).
- A decimal written into an `i` member went uncast, from inside the
  class (hb_csMemberIsInteger read declared types only) and from
  outside it (the scan registered member rows for declared types only;
  an undeclared `i` member is now an INTEGER row).
- The by-ref shim seeded a long temp from a decimal uncast
  (`Halve( @nX )` into an `i` parameter); the seed now casts, as the
  write-back already did.
- An `n` local typed INTEGER by its initializer - a call returning
  long, an integral #define, an AS INTEGER member - kept long when it
  later took a fraction, so `nMask := nMask / 2` truncated (1 in C#,
  1.50 in Harbour). The #define half predates the `i` work. Pass 2.5
  now sends such a local back to decimal on a hard disqualifier, and
  Pass 3 carries that down to the declaration. Nothing in either
  EasiPOS target was demoted: no such truncation was live.

Method parameters register their types for the body, as function
parameters already did; a method's `i` parameter read into a long had
kept a cast.

`this.` only where it disambiguates (Alex). A member of the class is
written bare unless a parameter, a local, a codeblock parameter in
scope (a stack pushed per codeblock), a PUBLIC or a MEMVAR has the
name; `this.Super()` and the other built-in object messages keep it
(extension methods need a receiver). INLINE bodies follow the same rule
on their text: any free use of the name, `INLINE ::cName := cName` for
one, keeps `this.`. easipos 1,467 -> 86, testeasipos 2,420 -> 84.

test121 (suite-unique method names: a shared `Count` had made
test103's OLE `oCom:Count` a call). Reference outputs: 27 files, every
changed line either `this.` removed, a redundant cast dropped (five)
or SetBit119 returning long. Suite 123/123, negative tests 18/18.

Both targets: scans 644/0/0 and 653/0/0, gens clean, builds 0 errors /
0 warnings. Against the last snapshot every changed line is `this.`
(1,178), a cast dropped (237), a setter returning long, or the
easiudm (decimal); dynamic unchanged at 14,823; audit TOTAL 2,021 ->
1,949 (easipos), 2,077 -> 2,005 (testeasipos).

+++ src/transpiler/tests/csout/test119.cs
@@ -27,7 +27,7 @@ public static partial class Program
         return;
     }
 
-    public static decimal SetBit119(decimal nFlags = default, bool lx = default)
+    public static long SetBit119(decimal nFlags = default, bool lx = default)
     {
         return (lx ? HbRuntime.hb_bitOr(nFlags, Test119PrgConst.FLAG119_BIT) : HbRuntime.hb_bitAnd(nFlags, HbRuntime.hb_bitNot(Test119PrgConst.FLAG119_BIT)));
     }
